Move ValidatesExistence into Concerns
